Added Events and classes for Context Subscription
Created TraCIContextSubscribableCommands that extends TraCICommandBase and is also an abstract class hat Classes that support Context Subscribe should extend.

Created VariableSubscribeEventArgs and ContextSubscribeEventArgs that exted SubscribeEventArgs for backwards compatability and for keeping old logic of working with subscriptions the same. With the new way Dictionaries can be used to get Variable Response by Variable code and for Context Subscribe a dictionary to get the values for the vehicle with objectId that was inside context range.

TraCIConverter changed to support context subscription.

Added events and event handler is ControlCommands for context subcsription.

Created interface IResponseInfo to help with value retrieval when you can't use TraCIResponse<T>. TraCIResponse<T> implements it.